pyproject.toml
[project]
name = "braindance"
version = "0.1.0"
description = "LeJEPA-EEG: Self-supervised pretraining for ENIGMA EEG encoder"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "torchvision>=0.16",
    "numpy>=1.24",
    "scipy>=1.11",
    "mne>=1.6",
    "lejepa",
    "open-clip-torch>=2.24",
    "einops>=0.7",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "matplotlib>=3.8",
    "scikit-learn>=1.3",
    "torchmetrics>=1.2",
    "wandb>=0.16",
    "pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff>=0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.sources]
lejepa = { path = "upstream/lejepa" }

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
